HW-3/requirements.txt
google-cloud-storage>=2.0.0
google-api-core>=2.0.0
google-cloud-bigquery